+++ bank-employees-salaries-impl/src/main/java/com/anybank/repository/EmployeeRepository.java
@@ -1,6 +1,9 @@
 package com.anybank.repository;
 
+
+import com.egorov.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+
 }
