+++ bank-employees-salaries-impl/src/main/java/com/anybank/repository/SalariesDateRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SalariesDateRepository extends JpaRepository<SalariesData, Integer> {
-    Optional<SalariesData> findByPosition(Integer id);
+public interface SalariesDateRepository extends JpaRepository<SalariesData, Long> {
+    Optional<SalariesData> findByPosition(Long id);
 }
