+++ bank-employees-salaries-impl/src/main/java/com/anybank/repository/SalaryRepository.java
@@ -7,13 +7,13 @@ import java.util.List;
 import java.util.Optional;
 
 public interface SalaryRepository extends JpaRepository<Salary, Long> {
-    Optional<Salary> findByEmployeeAndMonthAndYear(Integer employeeId, String month, String year);
+    Optional<Salary> findByEmployeeAndMonthAndYear(Long employeeId, String month, String year);
 
-    List<Salary> findByEmployeeAndYear(Integer employeeId, String year);
+    List<Salary> findByEmployeeAndYear(Long employeeId, String year);
 
-    List<Salary> findByDepartmentAndMonthAndYear(Integer departmentId, String month, String year);
+    List<Salary> findByDepartmentAndMonthAndYear(Long departmentId, String month, String year);
 
-    List<Salary> findByDepartmentAndYear(Integer departmentId, String year);
+    List<Salary> findByDepartmentAndYear(Long departmentId, String year);
 
     List<Salary> findByMonthAndYear(String month, String year);
